requirements.txt
fastapi
uvicorn
pandas
openpyxl
duckdb
sqlalchemy
psycopg2-binary
python-multipart
snf-peirce
tantivy
